+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# [0.86.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.85.0...v0.86.0) (2025-10-29)
+
+
+### Features
+
+* enhance enrollment settings with new fields and validation ([#672](https://github.com/dhis2/maintenance-app-beta/issues/672)) ([a80e16e](https://github.com/dhis2/maintenance-app-beta/commit/a80e16e3b85bbcbb19e5bc9cd352e5f00ce961ec))
+
+# [0.85.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.84.0...v0.85.0) (2025-10-28)
+
+
+### Features
+
+* program enrollment custom forms [DHIS2-20170] ([#673](https://github.com/dhis2/maintenance-app-beta/issues/673)) ([97632b8](https://github.com/dhis2/maintenance-app-beta/commit/97632b8e2f6df3e06c90e3e68914c223cbb6d13e))
+
+# [0.84.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.83.0...v0.84.0) (2025-10-28)
+
+
+### Features
+
+* programs list view ([#675](https://github.com/dhis2/maintenance-app-beta/issues/675)) ([b1c7d94](https://github.com/dhis2/maintenance-app-beta/commit/b1c7d941de8cd4026cc09b3f32a148c70e46814e))
+
+# [0.83.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.82.0...v0.83.0) (2025-10-28)
+
+
+### Features
+
+* add enrollment data section with fields ([#668](https://github.com/dhis2/maintenance-app-beta/issues/668)) ([561fc00](https://github.com/dhis2/maintenance-app-beta/commit/561fc009e43ec071088fca9dd28ce9852eab9612))
+
 # [0.82.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.81.1...v0.82.0) (2025-10-27)